chore: Refatorando Controller de Convenio

+++ GerenciadorHospital.Application/DTOs/Requests/ConvenioRequestDto.cs
@@ -3,5 +3,5 @@
 public class ConvenioRequestDto
 {
     public string Nome { get; set; } = string.Empty;
-    public decimal Preco { get; set; }
+    public float Preco { get; set; }
 }
